LPS-54903 Move JSP logic to BaseJSP class
